feat: separate client module, rename kiwi-talk-client to headless-talk
