+++ Cargo.toml
@@ -18,9 +18,9 @@ log = "0.4.20"
 anyhow = "1.0.75"
 serde = { version = "1.0", features = ["derive"] }
 
-[workspace.dependencies.kiwi-talk-client]
+[workspace.dependencies.headless-talk]
 version = "0.3.2"
-path = "./backend/crates/kiwi-talk-client"
+path = "./backend/crates/headless-talk"
 
 [workspace.dependencies.talk-api-internal]
 version = "0.1.2"
