+++ backend/bin/app/Cargo.toml
@@ -15,12 +15,12 @@ rust-version.workspace = true
 tauri-build = { version = "1.5.0", features = [] }
 
 [dependencies]
-kiwi-talk-client = { workspace = true }
 talk-loco-client = { workspace = true }
 
 kiwi-talk-system = { path = "../system" }
 kiwi-talk-result = { path = "../result" }
 kiwi-talk-api = { path = "../api" }
+kiwi-talk-client = { path = "../client" }
 
 tokio = { workspace = true, features = ["full"] }
 tauri = { workspace = true, features = [ "updater", "api-all", "system-tray"] }
@@ -29,18 +29,9 @@ log = { workspace = true }
 serde = { workspace = true }
 
 serde_json = "1.0"
-tokio-util = { version = "0.7.9", features = ["compat"] }
-thiserror = "1.0"
-once_cell = "1.18.0"
 env_logger = "0.10.0"
-parking_lot = "0.12.1"
-futures = "0.3.28"
 window-shadows = "0.2.1"
-num-bigint-dig = "0.8.4"
 tauri-plugin-window-state = { git = "https://github.com/tauri-apps/plugins-workspace", branch = "v1" }
-sha2 = "0.10.8"
-hex = "0.4.3"
-arrayvec = { version = "0.7.4", features = ["serde"] }
 
 [features]
 default = []
