+++ backend/bin/client/Cargo.toml
@@ -1,0 +1,34 @@
+[package]
+name = "kiwi-talk-client"
+description = "KiwiTalk client"
+publish = false
+
+version.workspace = true
+license.workspace = true
+edition.workspace = true
+authors.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[dependencies]
+talk-loco-client = { workspace = true }
+headless-talk = { workspace = true }
+
+kiwi-talk-api = { path = "../api" }
+kiwi-talk-system = { path = "../system" }
+kiwi-talk-result = { path = "../result" }
+
+tokio = { workspace = true, features = ["rt"] }
+tauri = { workspace = true }
+anyhow = { workspace = true }
+serde = { workspace = true }
+log = { workspace = true }
+
+futures = "0.3.28"
+sha2 = "0.10.8"
+parking_lot = "0.12.1"
+num-bigint-dig = "0.8.4"
+once_cell = "1.18.0"
+tokio-util = { version = "0.7.9", features = ["compat"] }
+arrayvec = "0.7.4"
+hex = "0.4.3"
