+++ backend/bin/system/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod constants;
+mod constants;
 
 use std::{
     ops::Deref,
