+++ backend/crates/headless-talk/src/channel/open.rs
