+++ backend/crates/headless-talk/src/chat/builder.rs
