+++ backend/crates/headless-talk/src/database/channel/open/model.rs
